Add files via upload
Challenge by Enterprise DNA
Challenge number 01
This Challenge was all about creating a summary report, a task that for many BI developers will always find its way across the desk every so often.
Overall the quality of submissions was super high, and you all did an amazing job at extrapolating what was required from a very vague brief.
It was great to see this coming through in the development through measures and visualisations.
When creating summary reports it’s so important that you are actually summarising data.
I think with all the visualisations available sometimes we overlook basics like tables and its important to remember that sometimes that’s all that is required. It’s all about understanding who your audience is.
So just want to say a massive well done to everyone who took part in the challenge and producing a management worthy dashboard.