Add files via upload
Challange 03 by Enterprise DNA
The Brief:
News is fast spreading about who the go to analyst is in the company. (PS it’s YOU)
The HR manager (James) has approached you for some help.
He is reviewing all internal consultations that have taken place around policy changes.
They currently have a HR system in place where they create and manage the respective consultations.
The current process is that once management decide that there is going to be a change to any policy, they then create a consultation within this HR system assigning a window to hold interviews with employees.
The company policy is that every employee will be interviewed twice regarding the change. The outcome of each interview is recorded.
Unfortunately the current system doesn’t have a reporting suite.
James has managed to get the data out of the system with some help from IT but isn’t sure where to go next.
He has said he is looking to report back to his seniors around the outcomes of these interviews and what the reception has been towards a consultation. James is also keen to establish the change of outcomes between the two interviews and any other metrics that you feel are key to this report.
The ball is now in your court and you need to find a way presenting this data back to James.