Add files via upload
Challange 04 - Enterprise DNA
The Brief
You are working at a consultancy that implemented an app for a client to help them improve their delivery process and fulfilment from warehouses to store.
The app was the created so that drivers could scan the label as opposed to entering it manually while also recording the time of arrival, the time they left the store and some other variables that the management were interested in improving. The management is now looking to evaluate how the business has reacted and how effective the app is. The senior consultant has extracted all data from the app and placed it into the file as below. He now wants you to use this data and create a report or dashboard to help management visualise what is happening so that we can decide what the next stage of the transformation is. The management are particularly interested in any warehouse store combination that it’s not working well for. The time being spent at a store.
The number of times manual entry was still used.
The number of damaged labels
If returns are being collected
A 1 in the data represents true
A 0 represents false
The ball is now in your court and you need to find a way presenting this data back to Janet and the team.